clean up integrations
