+++ flutter/lib/sentry_flutter.dart
@@ -1,7 +1,7 @@
 /// A Flutter client for Sentry.io crash reporting.
 export 'package:sentry/sentry.dart';
 
-export 'src/default_integrations.dart';
+export 'src/integrations/load_release_integration.dart';
 export 'src/navigation/sentry_navigator_observer.dart';
 export 'src/sentry_flutter.dart';
 export 'src/sentry_flutter_options.dart';
