+++ flutter/test/integrations/native_sdk_integration_test.dart
@@ -3,12 +3,13 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
+import 'package:sentry_flutter/src/integrations/native_sdk_integration.dart';
 import 'package:sentry_flutter/src/version.dart';
 
-import 'mocks.dart';
+import '../mocks.dart';
 
 void main() {
-  group('$NativeSdkIntegration', () {
+  group(NativeSdkIntegration, () {
     late Fixture fixture;
     setUp(() {
       fixture = Fixture();
