+++ flutter/test/integrations/not_initialized_widgets_binding_test.dart
@@ -1,7 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
+import 'package:sentry_flutter/src/integrations/widgets_binding_integration.dart';
 
-import 'mocks.mocks.dart';
+import '../mocks.mocks.dart';
 
 /// Tests that require `WidgetsFlutterBinding.ensureInitialized();` not
 /// being called at all.
